* Use a combination of loading attendance information with the page and ajax requests to change and refresh it. * Offer a single field only for substitution in JEvents layouts. * Field also usable in list layouts.

+++ ajax/simpleAttendance.php
@@ -10,28 +10,59 @@ $lang->load("plg_jevents_simple_attendance_ajax", JPATH_ADMINISTRATOR);
 
 class plgAjaxSimpleAttendance extends JPlugin
 {
+    function getAttendees($repetitionId) {        
+        $db    = JFactory::getDbo();
+        $query = $db->getQuery(true);            
+        $query->select('user_id') 
+                ->from($db->quoteName('#__simple_attendance'))
+                ->where(array('rp_id = '.(int)$repetitionId));
+        $db->setQuery($query);
+        $results = $db->loadObjectList();
+        
+        $attendees = array();
+        
+        foreach ($results as $result) {
+            $attendees[] = JFactory::getUser($result->user_id);
+        }
+        
+        return $attendees;
+    }   
+    
+    function getAttendanceInfo($repetitionId, $userId) {
+        $attendees = $this->getAttendees($repetitionId);
+        $getUserName = function($user) {return $user->username;};
+        $notMeFilter = function($otherUser) use($userId) {return $userId != $otherUser->id;};
+        
+        $result = new stdClass();
+        $result->otherAttendees = array_map($getUserName, array_filter($attendees, $notMeFilter));
+        $result->attendMyself = count($attendees) > count($result->otherAttendees);
+        return $result;
+    }
+    
     function onAjaxSimpleAttendance()
-    {        
+    {                
         $input        = JFactory::getApplication()->input;
         $repetitionId = $input->getInt('rp_id');
         $userId       = JFactory::getUser()->id;
-        $attend       = filter_var($input->get('attend', 'true'), FILTER_VALIDATE_BOOLEAN);
+        $attend       = is_null($input->get('attend')) ? null : filter_var($input->get('attend'), FILTER_VALIDATE_BOOLEAN);
         
-        //TODO: Permission check!
-        //TODO: Prevent SQL injection!     
-        $db    = JFactory::getDbo();
-        $query = $db->getQuery(true);
+        //TODO: Authorization check!                                      
+        if (isset($attend)) {
+            $db    = JFactory::getDbo();
+            $query = $db->getQuery(true);
+            if ($attend) {           
+                $query  ->insert($db->quoteName('#__simple_attendance'))
+                        ->columns($db->quoteName(array('rp_id', 'user_id')))
+                        ->values((int)$repetitionId . ',' . (int)$userId);        
+            } else {
+                $query ->delete($db->quoteName('#__simple_attendance'))
+                       ->where(array('rp_id = '. (int)$repetitionId, 'user_id = '.(int)$userId));
+            }
             
-        if ($attend) {           
-            $query  ->insert($db->quoteName('#__simple_attendance'))
-                    ->columns($db->quoteName(array('rp_id', 'user_id')))
-                    ->values($repetitionId . ',' . $userId);        
-        } else {
-            $query ->delete($db->quoteName('#__simple_attendance'))
-                   ->where(array('rp_id = '.$repetitionId, 'user_id = '.$userId));
+            $db->setQuery($query);
+            $db->execute();
         }
         
-        $db->setQuery($query);
-        return $db->execute();
+        return $this->getAttendanceInfo($repetitionId, $userId);
     }
 }
